Inserting a CMS Block right before the footer
Added a cms block using Targetables.
Identifiers for the cms block invented by me.
As a result, under the footer, the cms block is displayed with the content: 'The CMS block with the 'my-identifiers' ID doesn't exist.
As far as I understood, for correct content to be displayed, it needs to be added from admin area.